chore: use meterup/github-release to fix GitHub's deprecation.
```
[GitHub API] Deprecation notice for authentication via URL query parameters
----

On April 1st, 2020 at 23:12 (UTC) your personal access token (ruby-rpm > GITHUB_TOKEN in https://circleci.com/gh/feedforce/ruby-rpm/edit#env-vars) using Go-http-client/1.1 was used as part of a query parameter to access an endpoint through the GitHub API:

https://api.github.com/repositories/25909051/releases

Please use the Authorization HTTP header instead, as using the `access_token` query parameter is deprecated. If this token is being used by an app you don't have control over, be aware that it may stop working as a result of this deprecation.

Depending on your API usage, we'll be sending you this email reminder on a monthly basis for each token and User-Agent used in API calls made on your behalf.
Just one URL that was accessed with a token and User-Agent combination will be listed in the email reminder, not all.

Visit https://developer.github.com/changes/2020-02-10-deprecating-auth-through-query-param for more information about suggested workarounds and removal dates.

Thanks,
The GitHub Team
```